linux-bootloader: simplify get_loader_features
Rather than calling get_variable_size, allocating a vec, and reading it
into the buffer, use get_variable_boxed. Internally this does more or
less the same thing, so there's no change in behavior, just simplifies
things for the caller.

As before, it the variable cannot be read then the default features are
returned. If the data is the wrong size, an error is returned. Added
this behavior to the docstring for clarity.